Refactor for x layer
