+++ datastreamer/streamserver.go
@@ -50,7 +50,7 @@ const (
 	CmdStartBookmark                    // CmdStartBookmark for the start from bookmark TCP client command
 	CmdEntry                            // CmdEntry for the get entry TCP client command
 	CmdBookmark                         // CmdBookmark for the get bookmark TCP client command
-	CmdRangeBookmark                    // CmdRangeBookmark for the start and end bookmarks TCP client command
+	CmdRangeBookmark                    // For X Layer, CmdRangeBookmark for the start and end bookmarks TCP client command
 )
 
 const (
@@ -59,7 +59,7 @@ const (
 	CmdErrAlreadyStopped                      // CmdErrAlreadyStopped for client already stopped error
 	CmdErrBadFromEntry                        // CmdErrBadFromEntry for invalid starting entry number
 	CmdErrBadFromBookmark                     // CmdErrBadFromBookmark for invalid starting bookmark
-	CmdErrBadToBookmark                       // CmdErrBadToBookmark for invalid to bookmark
+	CmdErrBadToBookmark                       // For X Layer, CmdErrBadToBookmark for invalid to bookmark
 	CmdErrInvalidCommand  CommandError = 9    // CmdErrInvalidCommand for invalid/unknown command error
 )
 
@@ -96,7 +96,7 @@ var (
 		CmdStartBookmark: "StartBookmark",
 		CmdEntry:         "Entry",
 		CmdBookmark:      "Bookmark",
-		CmdRangeBookmark: "CmdRangeBookmark",
+		CmdRangeBookmark: "CmdRangeBookmark", // For X Layer
 	}
 
 	// StrCommandErrors for TCP command errors description
@@ -106,7 +106,7 @@ var (
 		CmdErrAlreadyStopped:  "Already stopped",
 		CmdErrBadFromEntry:    "Bad from entry",
 		CmdErrBadFromBookmark: "Bad from bookmark",
-		CmdErrBadToBookmark:   "Bad to bookmark",
+		CmdErrBadToBookmark:   "Bad to bookmark", // For X Layer
 		CmdErrInvalidCommand:  "Invalid command",
 	}
 )
@@ -784,6 +784,7 @@ func (s *StreamServer) processCommand(command Command, client *client) error {
 	case CmdBookmark:
 		err = s.handleBookmarkCommand(cli)
 
+	// For X Layer
 	case CmdRangeBookmark:
 		err = s.handleRangeBookmarkCommand(cli)
 
@@ -825,23 +826,6 @@ func (s *StreamServer) handleStartBookmarkCommand(cli *client) error {
 	err := s.processCmdStartBookmark(cli)
 	if err == nil {
 		cli.status = csSynced
-	}
-
-	return err
-}
-
-// handleRangeBookmarkCommand processes the CmdRangeBookmark command
-func (s *StreamServer) handleRangeBookmarkCommand(cli *client) error {
-	if cli.status != csStopped {
-		log.Error("Stream to client already started!")
-		_ = s.sendResultEntry(uint32(CmdErrAlreadyStarted), StrCommandErrors[CmdErrAlreadyStarted], cli)
-		return ErrClientAlreadyStarted
-	}
-
-	cli.status = csSyncing
-	err := s.processCmdRangeBookmark(cli)
-	if err == nil {
-		cli.status = csStopped
 	}
 
 	return err
@@ -972,83 +956,6 @@ func (s *StreamServer) processCmdStartBookmark(client *client) error {
 	}
 
 	return err
-}
-
-func (s *StreamServer) processCmdRangeBookmark(client *client) error {
-	// Read start and end bookmark parameter
-	sb, err := readBookmark(client)
-	if err != nil {
-		return err
-	}
-	eb, err := readBookmark(client)
-	if err != nil {
-		return err
-	}
-	log.Debugf("Client %s command RangeBookmark start: [%v], end [%v]", client.clientID, sb, eb)
-
-	from, err := s.bookmark.GetBookmark(sb)
-	if err != nil || from >= s.nextEntry {
-		log.Errorf("RangeBookmark command invalid start bookmark %v for client %s: %v", sb, client.clientID, err)
-		err = ErrStartBookmarkInvalidParamFromBookmark
-		_ = s.sendResultEntry(uint32(CmdErrBadFromBookmark), StrCommandErrors[CmdErrBadFromBookmark], client)
-		return err
-	}
-
-	to, err := s.bookmark.GetBookmark(eb)
-	if err != nil || to == 0 || to >= s.nextEntry {
-		log.Errorf("RangeBookmark command invalid end bookmark %v for client %s: %v", eb, client.clientID, err)
-		err = ErrEndBookmarkInvalidParamToBookmark
-		_ = s.sendResultEntry(uint32(CmdErrBadToBookmark), StrCommandErrors[CmdErrBadToBookmark], client)
-		return err
-	}
-
-	// Find L2BlockEnd or BatchEnd entry number
-	iterator, err := s.streamFile.iteratorFrom(to, true)
-	if err != nil {
-		log.Errorf("RangeBookmark command invalid to iterator %v for client %s: %v", eb, client.clientID, err)
-		err = ErrEndBookmarkInvalidParamToBookmark
-		_ = s.sendResultEntry(uint32(CmdErrBadToBookmark), StrCommandErrors[CmdErrBadToBookmark], client)
-		return err
-	}
-	reachedBlockEnd := false
-	for {
-		// Get next entry data
-		end, err := s.streamFile.iteratorNext(iterator)
-		if err != nil {
-			log.Warnf("RangeBookmark command invalid next iterator %v for client %s: %v", eb, client.clientID, err)
-			err = ErrEndBookmarkInvalidParamToBookmark
-			_ = s.sendResultEntry(uint32(CmdErrBadToBookmark), StrCommandErrors[CmdErrBadToBookmark], client)
-			return err
-		}
-
-		// Check if end of iterator or reached end of the range
-		if end || (reachedBlockEnd && iterator.Entry.Type != EtBatchEnd) {
-			break
-		}
-
-		if iterator.Entry.Type == EtL2BlockEnd {
-			// Reached the end of a block. Continue iteration to check for batch end
-			reachedBlockEnd = true
-			to = iterator.Entry.Number
-		} else if iterator.Entry.Type == EtBatchEnd {
-			// Reached the end of a batch
-			to = iterator.Entry.Number
-			break
-		}
-	}
-
-	// Send a command result entry OK
-	err = s.sendResultEntry(0, "OK", client)
-	if err != nil {
-		return err
-	}
-
-	// Send toEntry
-	be := make([]byte, 8)
-	binary.BigEndian.PutUint64(be, to)
-	TimeoutWrite(client, be, s.writeTimeout)
-
-	return s.streamingRangeEntry(client, from, to)
 }
 
 // processCmdStop processes the TCP Stop command from the clients
@@ -1230,58 +1137,6 @@ func (s *StreamServer) streamingFromEntry(client *client, fromEntry uint64) erro
 	return nil
 }
 
-// streamingRangeEntry streams the range of file entries from entry number until to entry number
-func (s *StreamServer) streamingRangeEntry(client *client, fromEntry uint64, toEntry uint64) error {
-	if fromEntry > toEntry {
-		return ErrInvalidBookmarkRange
-	}
-
-	log.Debugf("SYNCING %s from entry %d to entry %d...", client.clientID, fromEntry, toEntry)
-
-	// Start file stream iterator
-	iterator, err := s.streamFile.iteratorFrom(fromEntry, true)
-	if err != nil {
-		return err
-	}
-
-	// Loop data entries from file stream iterator
-	for {
-		// Get next entry data
-		end, err := s.streamFile.iteratorNext(iterator)
-		if err != nil {
-			return err
-		}
-
-		// Check if end of iterator
-		if end {
-			break
-		}
-
-		// Send the file data entry
-		binaryEntry := encodeFileEntryToBinary(iterator.Entry)
-		log.Debugf("Sending data entry %d (type %d) to %s", iterator.Entry.Number, iterator.Entry.Type, client.clientID)
-		if client.conn != nil {
-			_, err = TimeoutWrite(client, binaryEntry, s.writeTimeout)
-		} else {
-			err = ErrNilConnection
-		}
-		if err != nil {
-			log.Errorf("Error sending entry %d to %s: %v", iterator.Entry.Number, client.clientID, err)
-			return err
-		}
-
-		if iterator.Entry.Number == toEntry {
-			break
-		}
-	}
-	log.Debugf("Synced %s until %d!", client.clientID, iterator.Entry.Number)
-
-	// Close iterator
-	s.streamFile.iteratorEnd(iterator)
-
-	return nil
-}
-
 // sendResultEntry sends the response to a TCP command for the clients
 func (s *StreamServer) sendResultEntry(errorNum uint32, errorStr string, client *client) error {
 	// Prepare the result entry
@@ -1330,29 +1185,6 @@ func (s *StreamServer) BookmarkPrintDump() {
 	if err != nil {
 		log.Errorf("Error dumping bookmark database")
 	}
-}
-
-func readBookmark(client *client) ([]byte, error) {
-	// Read bookmark length parameter
-	length, err := readFullUint32(client)
-	if err != nil {
-		return nil, err
-	}
-
-	// Check maximum length allowed
-	if length > maxBookmarkLength {
-		log.Errorf("Client %s exceeded [%d] maximum allowed length [%d] for a bookmark.",
-			client.clientID, length, maxBookmarkLength)
-		return nil, ErrBookmarkMaxLength
-	}
-
-	// Read start bookmark parameter
-	bookmark, err := readFullBytes(length, client)
-	if err != nil {
-		return nil, err
-	}
-
-	return bookmark, nil
 }
 
 // readFullUint64 reads from a connection a complete uint64
